test: split tests that differ from v0
